Update Django Final

+++ postings/views.py
@@ -2,7 +2,7 @@ from django.shortcuts import render, redirect, get_object_or_404
 from .models import Posting, Comment
 from .forms import PostingForm, CommentForm
 # [코드 작성] django.contrib.auth.decorators 에서 login_required 데코레이션 가져오기
-from django.contrib.auth.decorators import login_required
+
 # [코드 작성] django.views.decorators.http 에서 require_POST 데코레이션 가져오기
 
 # Paginator 가져오기
@@ -81,7 +81,8 @@ def posting_detail(request, posting_id):
     }
     return render(request, 'postings/posting_detail.html', context)
 
-@login_required
+# [코드 작성] login_required 데코레이션 추가
+
 def posting_update(request, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
     if posting.author == request.user:
@@ -103,25 +104,27 @@ def posting_update(request, posting_id):
     return redirect('postings:posting_detail', posting_id)
 
 # [코드 작성] login_required 데코레이션 추가
-@login_required
+
 # [코드 작성] require_POST 데코레이션 추가
 
 def posting_delete(request, posting_id):
     posting = get_object_or_404(Posting, id=posting_id)
-    # [코드 작성] 글(posting) 작성자(author)가 로그인한 사람(request.user)과 같을 경우에만 글 삭제가 가능하도록 조건문 작성
-    
-    posting.delete()
-    return redirect('postings:posting_list')
-    # [코드 작성] posting_id에 해당하는 페이지로 redirect
-    
+    # [미션] 글(posting) 작성자(author)가 로그인한 사람(request.user)과 같을 경우에만 글 삭제가 가능하도록 조건문 작성
+    # [미션] True를 지우고 작성
+    if True:
+        posting.delete()
+        return redirect('postings:posting_list')
+    # [미션] posting_id에 해당하는 페이지로 redirect
+    return 
 
 # [코드 작성] login_required 데코레이션 추가
-@login_required
+
 # [코드 작성] require_POST 데코레이션 추가
 
 def comment_delete(request, posting_id, comment_id):
     comment = get_object_or_404(Comment, id=comment_id)
-    # [코드 작성] 댓글(comment) 작성자(author)가 로그인한 사람(request.user)과 같을 경우에만 댓글 삭제가 가능하도록 조건문 작성
-    
-    comment.delete()
+    # [미션] 댓글(comment) 작성자(author)가 로그인한 사람(request.user)과 같을 경우에만 댓글 삭제가 가능하도록 조건문 작성
+    # [미션] True를 지우고 작성
+    if True:
+        comment.delete()
     return redirect('postings:posting_detail', posting_id)
